Fix arithmetic_side_effects: display, shift, inv, pi, refinement
- display.rs: convert num_bits to BigInt before subtraction
- shift.rs: add ArbitraryPrecision marker trait, narrow #[allow] to <<= line
- inv.rs: use saturating_neg() for precision estimate
- pi.rs: assert_sane_computation_size! guards, named variables, saturating_mul
- refinement.rs: replace manual counters with for loops

+++ src/binary/display.rs
@@ -65,7 +65,7 @@ fn format_mantissa_with_point(mantissa: &BigUint, exponent: &BigInt) -> (String,
     };
 
     // Adjust exponent: new_exp = old_exp + (num_bits - 1)
-    let adjusted_exponent = exponent + (num_bits - 1);
+    let adjusted_exponent = exponent + BigInt::from(num_bits) - 1_i32;
 
     (formatted_mantissa, adjusted_exponent)
 }
